+++ spotbugs/src/main/java/edu/umd/cs/findbugs/ba/ValueRangeDataflow.java
@@ -1,0 +1,8 @@
+package edu.umd.cs.findbugs.ba;
+
+public class ValueRangeDataflow extends Dataflow<ValueRangeMap, ValueRangeAnalysis> {
+
+    public ValueRangeDataflow(CFG cfg, ValueRangeAnalysis analysis) {
+        super(cfg, analysis);
+    }
+}
